// master/client/assets/script/game/entity/FriendInfo.ts
export type FriendInfo = {
    /** 角色头像 */
    roleIcon: number;
    /** 角色名字 */
    roleName: string;
    /** 角色id */
    roleId: number;
    /** 角色战斗力 */
    battlePoint: number;
    /** 角色等级 */
    level: number;
    /** 角色最后上线时间 */
    lastOnLineTime: number;
    /** 角色头像框 */
    headFrame: number;
};